notifications/event-onclick-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - onclick (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="interact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/notifications.js"></script>
<body>
  <p>Test steps:</p>
    <ol>
      <li>When a notification popups, click the notification's body.</li>
    </ol>
  <div id="log"></div>
  <script>
    t = async_test("the notification.onclick event can be invoked", {timeout: 50000});
    getnotification();
    notification.show();
    notification.onclick = t.step_func(function(e) {
      assert_equals(Object.prototype.toString.call(e), "[object Event]", "the type of event");
      t.done();
    });
  </script>
</body>
